one makeTables to build them all
