+++ index.js
@@ -63,60 +63,6 @@ sequelize
 });
 // ****** Set up default MYSQL connection END ****** //
 
-// Set up all the machinery needed to produce a USER table
-// User model for sequelize
-const User = sequelize.define('user', {
-  first: { type: Sequelize.STRING },
-  last: { type: Sequelize.STRING },
-  username: { type: Sequelize.STRING },
-  email: { type: Sequelize.STRING },
-  password: { type: Sequelize.STRING },
-  roles: { type: Sequelize.STRING }
-},
-{
-  charset: 'utf8mb4',
-  collate: 'utf8mb4_unicode_ci'
-});
-
-// build the users table
-async function makeUsersTable(){
-// force: true will drop the table if it already exists
-await User.sync({force: true})
-.then(() => {
-  // Table created
-  return User.create({
-    first: "Original",
-    last: "Data",
-    username: "original",
-    email: 'colrc@gmail.com',
-    password: 'colrc@gmail.com',
-    roles: "admin"
-  });
-})
-.then((user) => {
-  console.log(user);
-  return User.findOne({
-    where: { id: 1 }
-  }).then((res) => {
-    return [{
-      id: res.dataValues.id,
-      first: res.dataValues.first,
-      last: res.dataValues.last,
-      username: res.dataValues.username,
-      password: res.dataValues.password,
-      email: res.dataValues.email,
-      roles: res.dataValues.roles.split(',')
-    }];
-  });
-})
-.then((newuser) => {
-  console.log(newuser);
-  console.log("COLRC");
-});
-}
-
-// next, build the Root Dictionary, Affix List and Stem List
-// define the entries for Root Dictionary
 const Root = sequelize.define('root', {
   root: { type: Sequelize.STRING },
   number: { type: Sequelize.INTEGER },
@@ -170,89 +116,28 @@ const Stem = sequelize.define('stem', {
   collate: 'utf8mb4_unicode_ci'
 });
 
-// build the roots table by reading in entries from the entries.txt file, which is ::: delimited
-// be sure to customize the path to find the file on your machine
-async function makeRootTable(){
-  await Root.sync({force: true});
-  var fs = require('fs');
-  var contents = fs. readFileSync('/Users/angel/Documents/src/data_files/entries.txt', 'utf8');
-  var rows = contents.split("\n");
-  rows.forEach(async function (row, index) {
-    columns = row.split(":::");
-    await Root.create({
-      root: columns[2],
-      number: parseInt(columns[3]),
-      salish: columns[4],
-      nicodemus: columns[5],
-      english: columns[6],
-      editnote: Sequelize.NULL,
-      active: 'Y',
-      prevId: Sequelize.NULL,
-      userId: "1"
-    });
-  });
-  console.log("I have a roots table");
-}
+const User = sequelize.define('user', {
+  first: { type: Sequelize.STRING },
+  last: { type: Sequelize.STRING },
+  username: { type: Sequelize.STRING },
+  email: { type: Sequelize.STRING },
+  password: { type: Sequelize.STRING },
+  roles: { type: Sequelize.STRING }
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
 
-// build the affixes table by reading in entries from the affixes.txt file, which is ::: delimited
-// be sure to customize the path to find the file on your machine
-async function makeAffixTable(){
-  await Affix.sync({force: true});
-  var fs = require('fs');
-  var contents = fs. readFileSync('/Users/angel/Documents/src/data_files/affixes.txt', 'utf8');
-  var rows = contents.split("\n");
-  rows.forEach(async function (row, index) {
-    columns = row.split(":::");
-    await Affix.create({
-      type: columns[0],
-      salish: columns[1],
-      nicodemus: columns[2],
-      english: columns[3],
-      link: columns[4],
-      page: columns[5],
-      editnote: Sequelize.NULL,
-      active: 'Y',
-      prevId: Sequelize.NULL,
-      userId: "1"
-    });
-  });
-  console.log("I have an affixes table");
-}
-// build the stems table by reading in entries from the stems_both_lists.txt file, which is ::: delimited.
-// be sure to customize the path to find the file on your machine
-async function makeStemTable(){
-  await Stem.sync({force: true});
-  var fs = require('fs');
-  var contents = fs. readFileSync('/Users/angel/Documents/src/data_files/stems_both_lists.txt', 'utf8');
-  var rows = contents.split("\n");
-  rows.forEach(async function (row, index) {
-    columns = row.split(":::");
-    await Stem.create({
-      category: columns[0],
-      reichard: columns[2],
-      doak: columns[3],
-      salish: columns[4],
-      nicodemus: columns[5],
-      english: columns[6],
-      note: columns[7],
-      editnote: Sequelize.NULL,
-      active: 'Y',
-      prevId: Sequelize.NULL,
-      userId: "1"
-    });
-  });
-  console.log("I have a stems table");
-}
+// build the tables required for the 'texts', 'audio' and 'elicitations' areas.
+// we need the following tables:  'texts', 'textfiles', 'textimages', 'audiosets' and 'audiofiles'
 
-// Now build the bibliography table
-// Define the type for bibliography
-const Bibliography = sequelize.define('bibliography', {
-  author: { type: Sequelize.STRING },
-  year: { type: Sequelize.STRING },
+// first, we need a text type
+
+const Text = sequelize.define('text', {
   title: { type: Sequelize.STRING },
-  reference: { type: Sequelize.STRING },
-  link: { type: Sequelize.STRING },
-  linktext: { type: Sequelize.STRING },
+  speaker: { type: Sequelize.STRING },
+  cycle: { type: Sequelize.STRING },
   active: { type: Sequelize.STRING(1) },
   prevId: { type: Sequelize.INTEGER },
   userId: { type: Sequelize.STRING }
@@ -262,29 +147,136 @@ const Bibliography = sequelize.define('bibliography', {
   collate: 'utf8mb4_unicode_ci'
 });
 
-// make the bibliography table, using Data.js
-async function makeBibliographyTable(){
-	await Bibliography.sync({force: true});
-	var contents = data.bibliography;
-	contents.forEach(async function (row) {
-		await Bibliography.create({
-      author: row.author,
-      year: row.year,
-      title: row.title,
-      reference: row.reference,
-      link: row.link,
-      linktext: row.linktext,
-      active: 'Y',
-      prevId: Sequelize.NULL,
-      userId: "1"
-		});
-	});
-	console.log("I have a bibliography table");
-}
+// now we need a textfiles type
+const Textfile = sequelize.define('textfile', {
+  subdir: { type: Sequelize.STRING },
+  src: { type: Sequelize.STRING },
+  resType: { type: Sequelize.STRING },
+  msType: { type: Sequelize.STRING },
+  fileType: { type: Sequelize.STRING },
+  textID: { type: Sequelize.STRING },
+  active: { type: Sequelize.STRING(1) },
+  prevId: { type: Sequelize.INTEGER },
+  userId: { type: Sequelize.STRING }
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
+
+// set up the Text to Textfile relationship
+const Texttofilerelation = sequelize.define('texttofilerelation', {
+  textId: { type: Sequelize.STRING },
+  textfileId: { type: Sequelize.STRING },
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
+
+// now we need a textimages type
+const Textimage = sequelize.define('textimage', {
+  subdir: { type: Sequelize.STRING },
+  src: { type: Sequelize.STRING },
+  active: { type: Sequelize.STRING(1) },
+  prevId: { type: Sequelize.INTEGER },
+  userId: { type: Sequelize.STRING }
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
+
+// set up the Text to Textfile relationship
+const Filetoimagerelation = sequelize.define('filetoimagerelation', {
+  textfileId: { type: Sequelize.STRING },
+  textimageId: { type: Sequelize.STRING },
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
+
+// next let's get an audiofile type
+const Audiofile = sequelize.define('audiofile', {
+  subdir: { type: Sequelize.STRING },
+  src: { type: Sequelize.STRING },
+  type: { type: Sequelize.STRING },
+  direct: { type: Sequelize.STRING },
+  active: { type: Sequelize.STRING(1) },
+  userId: { type: Sequelize.STRING }
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
+
+// then an audioset type
+const Audioset = sequelize.define('audioset', {
+  title: { type: Sequelize.STRING },
+  speaker: { type: Sequelize.STRING },
+  active: { type: Sequelize.STRING(1) },
+  textId: { type: Sequelize.STRING },
+  userId: { type: Sequelize.STRING }
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
+
+const Audiorelation = sequelize.define('audiorelation', {
+  audiosetId: { type: Sequelize.STRING },
+  audiofileId: { type: Sequelize.STRING },
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
+
+// set up the Text to Audioset relationship
+const Texttoaudiosetrelation = sequelize.define('texttoaudiosetrelation', {
+  textId: { type: Sequelize.STRING },
+  audiosetId: { type: Sequelize.STRING },
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
 
 
-// all the machinery needed to create the Spelling and Pronunciation Guide materials
-// first, type definitions
+const Elicitationset = sequelize.define('elicitationset', {
+  title: { type: Sequelize.STRING },
+  active: { type: Sequelize.STRING(1) },
+  userId: { type: Sequelize.STRING },
+  prevID: { type: Sequelize.INTEGER }
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
+
+const Elicitationfile = sequelize.define('elicitationfile', {
+  src: { type: Sequelize.STRING },
+  type: { type: Sequelize.STRING },
+  direct: { type: Sequelize.STRING },
+  active: { type: Sequelize.STRING(1) },
+  userId: { type: Sequelize.STRING }
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
+
+// then an elicitations to elicitationfile relation
+const Elicitationrelation = sequelize.define('elicitationrelation', {
+  elicitationsetId: { type: Sequelize.STRING },
+  elicitationfileId: { type: Sequelize.STRING },
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
+
 // this one builds the spelling list
 const Spelling = sequelize.define('spelling', {
   reichard: { type: Sequelize.STRING },
@@ -334,8 +326,160 @@ const Consonant = sequelize.define('consonant', {
   charset: 'utf8mb4',
   collate: 'utf8mb4_unicode_ci'
 });
+// Define the type for bibliography
+const Bibliography = sequelize.define('bibliography', {
+  author: { type: Sequelize.STRING },
+  year: { type: Sequelize.STRING },
+  title: { type: Sequelize.STRING },
+  reference: { type: Sequelize.STRING },
+  link: { type: Sequelize.STRING },
+  linktext: { type: Sequelize.STRING },
+  active: { type: Sequelize.STRING(1) },
+  prevId: { type: Sequelize.INTEGER },
+  userId: { type: Sequelize.STRING }
+},
+{
+  charset: 'utf8mb4',
+  collate: 'utf8mb4_unicode_ci'
+});
 
-// now build the tables
+// build the users table
+async function makeUsersTable(){
+// force: true will drop the table if it already exists
+await User.sync({force: true})
+.then(() => {
+  // Table created
+  return User.create({
+    first: "Original",
+    last: "Data",
+    username: "original",
+    email: 'colrc@gmail.com',
+    password: 'colrc@gmail.com',
+    roles: "admin"
+  });
+})
+.then((user) => {
+  console.log(user);
+  return User.findOne({
+    where: { id: 1 }
+  }).then((res) => {
+    return [{
+      id: res.dataValues.id,
+      first: res.dataValues.first,
+      last: res.dataValues.last,
+      username: res.dataValues.username,
+      password: res.dataValues.password,
+      email: res.dataValues.email,
+      roles: res.dataValues.roles.split(',')
+    }];
+  });
+})
+.then((newuser) => {
+  console.log(newuser);
+  console.log("COLRC");
+});
+}
+
+// next, build the Root Dictionary, Affix List and Stem List
+// build the roots table by reading in entries from the entries.txt file, which is ::: delimited
+// be sure to customize the path to find the file on your machine
+async function makeRootTable(){
+  await Root.sync({force: true});
+  var fs = require('fs');
+  var contents = fs. readFileSync('data\\entries.txt', 'utf8');
+  var rows = contents.split("\n");
+  rows.forEach(async function (row, index) {
+    columns = row.split(":::");
+    if (columns[5]) {
+      await Root.create({
+        root: columns[2],
+        number: parseInt(columns[3]),
+        salish: columns[4],
+        nicodemus: columns[5],
+        english: columns[6],
+        editnote: Sequelize.NULL,
+        active: 'Y',
+        prevId: Sequelize.NULL,
+        userId: "1"
+      });
+    }
+  });
+  console.log("I have a roots table");
+}
+// build the affixes table by reading in entries from the affixes.txt file, which is ::: delimited
+// be sure to customize the path to find the file on your machine
+async function makeAffixTable(){
+  await Affix.sync({force: true});
+  var fs = require('fs');
+  var contents = fs. readFileSync('data\\affixes.txt', 'utf8');
+  var rows = contents.split("\n");
+  rows.forEach(async function (row, index) {
+    columns = row.split(":::");
+    if (columns[2]) {    
+      await Affix.create({
+        type: columns[0],
+        salish: columns[1],
+        nicodemus: columns[2],
+        english: columns[3],
+        link: columns[4],
+        page: columns[5],
+        editnote: Sequelize.NULL,
+        active: 'Y',
+        prevId: Sequelize.NULL,
+        userId: "1"
+      });
+    }
+  });
+  console.log("I have an affixes table");
+}
+// build the stems table by reading in entries from the stems_both_lists.txt file, which is ::: delimited.
+// be sure to customize the path to find the file on your machine
+async function makeStemTable(){
+  await Stem.sync({force: true});
+  var fs = require('fs');
+  var contents = fs. readFileSync('data\\stems_both_lists.txt', 'utf8');
+  var rows = contents.split("\n");
+  rows.forEach(async function (row, index) {
+    columns = row.split(":::");
+    if (columns[5]) {
+      await Stem.create({
+        category: columns[0],
+        reichard: columns[2],
+        doak: columns[3],
+        salish: columns[4],
+        nicodemus: columns[5],
+        english: columns[6],
+        note: columns[7],
+        editnote: Sequelize.NULL,
+        active: 'Y',
+        prevId: Sequelize.NULL,
+        userId: "1"
+      });
+    }
+  });
+  console.log("I have a stems table");
+}
+
+// make the bibliography table, using Data.js
+async function makeBibliographyTable(){
+	await Bibliography.sync({force: true});
+	var contents = data.bibliography;
+	contents.forEach(async function (row) {
+		await Bibliography.create({
+      author: row.author,
+      year: row.year,
+      title: row.title,
+      reference: row.reference,
+      link: row.link,
+      linktext: row.linktext,
+      active: 'Y',
+      prevId: Sequelize.NULL,
+      userId: "1"
+		});
+	});
+	console.log("I have a bibliography table");
+}
+
 // this table builds the spelling list, using Data.js
 async function makeSpellingTable(){
   await Spelling.sync({force: true});
@@ -391,25 +535,238 @@ async function makeVowelTable(){
   });
   console.log("I have a vowel table");
 }
+//make all the relation tables
+//make the filetoimagerelation table
+async function makeFiletoimagerelationTable(){
+  // Define the relationships, which will automatically create a table
+  await Filetoimagerelation.sync({force: true});
+}
+//make the texttofilerelation table
+async function makeTexttofilerelationTable(){
+  // Define the relationships, which will automatically create a table
+  await Texttofilerelation.sync({force: true});
+}
+//make the texttoaudiosetrelation table
+async function makeTexttoaudiosetrelationTable(){
+  // Define the relationships, which will automatically create a table
+  await Texttoaudiosetrelation.sync({force: true});
+}
+// make the audiorelation table
+async function makeAudiorelationTable(){
+  // Define the relationships, which will automatically create a table
+  await Audiorelation.sync({force: true});
+}
+//make the elicitationrelation table
+async function makeElicitationrelationTable(){
+  await Elicitationrelation.sync({force: true});
+}
+
+//Set up all the relationships as foreign keys
+//Textfiles are linked to textimages
+Textfile.belongsToMany( Textimage, {
+  through: "filetoimagerelations", //this can be string or a model,
+  foreignKey: 'TextfileId'
+})
+Textimage.belongsToMany( Textfile, {
+  through: "filetoimagerelations", //this can be string or a model,
+  foreignKey: 'TextimageId'
+})
+//Texts are linked to textfiles
+Text.belongsToMany( Textfile, {
+  through: "texttofilerelations", //this can be string or a model,
+  foreignKey: 'TextId'
+})
+Textfile.belongsToMany( Text, {
+  through: "texttofilerelations", //this can be string or a model,
+  foreignKey: 'TextfileId'
+})
+//Texts are linked to audiosets
+Text.belongsToMany( Audioset, {
+  through: "texttoaudiosetrelations", //this can be string or a model,
+  foreignKey: 'TextId'
+})
+Audioset.belongsToMany( Text, {
+  through: "texttoaudiosetrelations", //this can be string or a model,
+  foreignKey: 'AudiosetId'
+})
+//audiosets are linked to audiofiles
+Audioset.belongsToMany( Audiofile, {
+  //as: [SetToFile],
+  through: "audiorelations", //this can be string or a model,
+  foreignKey: 'AudiosetId'
+})
+Audiofile.belongsToMany( Audioset, {
+  //as: [FileToSet],
+  through: "audiorelations",
+  foreignKey: 'AudiofileId'
+})
+
+
+//individual table builders, for the record.
+// make the Textimages table, using Data.js
+async function makeTextimageTable(){
+  await makeFiletoimagerelationTable();
+  await makeTextfileTable();
+  await Textimage.sync({force: true});
+  data.textimages.forEach(async function (row) {
+    let newTextImage = await Textimage.create({
+      subdir: row.subdir,
+      src: row.src,
+      active: 'Y',
+      prevId: Sequelize.NULL,
+      userId: '1'
+    })
+    let myTextFile = await Textfile.findOne({  where: {id: row.textfileId} })
+    await newTextImage.addTextfile(myTextFile)    
+  });
+  console.log("I have a textimages table");
+}
+
+// make the Textfiles table, using Data.js
+async function makeTextfileTable(){
+  await makeTexttofilerelationTable(); 
+  await Textfile.sync({force: true});
+  data.textfiles.forEach(async function (row) {
+    let newTextfile = await Textfile.create({
+      subdir: row.subdir,
+      src: row.src,
+      resType: row.resType,
+      msType: row.msType,
+      fileType: row.fileType,
+      textID: row.textId,
+      active: 'Y',
+      prevId: Sequelize.NULL,
+      userId: '1'
+    })
+    let myText = await Text.findOne({  where: {id: row.textId} })
+    await newTextfile.addText(myText)  
+  });
+  console.log("I have a textfiles table")
+}
+
+// make the Text table, using Data.js
+async function makeTextTable(){
+  await Text.sync({force: true});
+  data.texts.forEach(async function (row) {
+    await Text.create({
+      title: row.title,
+      speaker: row.speaker,
+      cycle: row.cycle,
+      active: 'Y',
+      prevId: Sequelize.NULL,
+      userId: '1'
+    });
+  });
+  console.log("I have a texts table");
+}
+
+// make the audioset table
+async function makeAudiosetTable(){
+  await makeTexttoaudiosetrelationTable();  
+  await Audioset.sync({force: true});
+  data.audiosets.forEach(async function (row) {
+    let newAudioSet = await Audioset.create({
+      title: row.title,
+      speaker: row.speaker,
+      textId: row.textId,
+      active: 'Y',
+      prevId: Sequelize.NULL,
+      userId: '1'
+    })
+    let myText = await Text.findOne({  where: {id: row.textId} })
+    await newAudioSet.addText(myText)
+  });
+  console.log("I have an audiosets table");
+}
+
+// make the audiofile table from Data.js
+async function makeAudiofileTable(){
+  await makeAudiosetTable()
+  await makeAudiorelationTable()
+  await Audiofile.sync({force: true})
+  data.audiofiles.forEach(await async function (row) {
+    let newAudioFile = await Audiofile.create({
+      subdir: row.subdir,
+      src: row.src,
+      type: row.type,
+      direct: row.direct,
+      active: 'Y',
+      prevId: Sequelize.NULL,
+      userId: '1'
+    })
+    let myAudioSet = await Audioset.findOne({  where: {id: row.audiosetId} })
+    await newAudioFile.addAudioset(myAudioSet)
+  })
+  console.log("I have an audiofiles table")
+}
+// Elicitations aren't linked to texts, so they're here.
+//elicitationsets are linked to elicitationfiles
+Elicitationset.belongsToMany( Elicitationfile, {
+  through: "elicitationrelations", //this can be string or a model,
+  foreignKey: 'elicitationsetId'
+})
+Elicitationfile.belongsToMany( Elicitationset, {
+  through: "elicitationrelations",
+  foreignKey: 'elicitationfileId'
+})
+// make the elicitationset table
+async function makeElicitationsetTable(){
+  await Elicitationset.sync({force: true});
+  data.elicitationsets.forEach(async function (row) {
+    await Elicitationset.create({
+      title: row.title,
+      active: 'Y',
+      prevId: Sequelize.NULL,
+      userId: '1'
+    });
+  });
+  console.log("I have an elicitationsets table");
+}
+
+// make the elicitationfile table from Data.js
+async function makeElicitationfileTable(){
+  await makeElicitationrelationTable();
+  await makeElicitationsetTable();
+  await Elicitationfile.sync({force: true});
+  data.elicitationfiles.forEach(async function (row) {
+    newElicitationFile = await Elicitationfile.create({
+      src: row.src,
+      type: row.type,
+      direct: row.direct,
+      active: 'Y',
+      prevId: Sequelize.NULL,
+      userId: '1'
+    })
+    let myElicitationSet = await Elicitationset.findOne({  where: {id: row.elicitationId} })
+    await newElicitationFile.addElicitationset(myElicitationSet)
+  });
+  console.log("I have an elicitationfiles table");
+}
+
+async function makeMedia(){
+  await makeFiletoimagerelationTable();
+  await makeTexttofilerelationTable();
+  await makeTexttoaudiosetrelationTable();
+  await makeAudiorelationTable();
+  await makeTextTable();    
+  await makeTextimageTable(); 
+  await makeAudiofileTable();
+  await makeElicitationfileTable();
+}
+
+async function makeTables(){
+  await makeUsersTable();
+  await makeRootTable();
+  await makeStemTable();
+  await makeAffixTable();
+  await makeUsersTable();
+  await makeSpellingTable();
+  await makeConsonantTable();
+  await makeVowelTable();
+  await makeMedia();
+}
 
 //  all function calls to build tables are below.  Uncomment the ones you want to build
 
-makeSpellingTable();
+makeTables();
 
-// makeUsersTable();
-
-// makeAffixTable();
-
-// makeRootTable();
-
-// makeStemTable();
-
-
-//makeBibliographyTable();
-
-// makeVowelTable();
-
-makeConsonantTable();
-
-// app.use('/', (req, res) => res.send("Welcome COLRC User"));
-// app.listen(process.env.GRAPHQLPORT, () => console.log('COLRC Enterprise Server is ready on localhost:' + process.env.GRAPHQLPORT));
